+++ documentation/ChapterArticle.json
@@ -1,25 +1,25 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
-  "owner" : "1yZ5ydzIh4dK",
+  "status" : "DRAFT",
+  "owner" : "kIQAH5a1slh0NBfR",
   "resourceOwner" : {
-    "owner" : "1yZ5ydzIh4dK",
-    "ownerAffiliation" : "https://www.example.org/voluptatibusfuga"
+    "owner" : "kIQAH5a1slh0NBfR",
+    "ownerAffiliation" : "https://www.example.org/consequaturquos"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/cumqueet",
+    "id" : "https://www.example.org/suntperspiciatis",
     "labels" : {
-      "UdDH70cuNGmqgf7Ck4" : "kKj5lNrdXC8w9K58i9X"
+      "bfne2q63Sdk1U" : "kmH5O2IAZ4Dobrg"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/minimaeos",
-  "doi" : "https://doi.org/10.1234/omnis",
+  "handle" : "https://www.example.org/ametvoluptatem",
+  "doi" : "https://doi.org/10.1234/numquam",
   "doiRequest" : {
     "type" : "DoiRequest",
     "status" : "APPROVED",
@@ -27,127 +27,127 @@
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "GMHhlSFyUMwjH1rlnE",
-      "author" : "I0dmBio3QZ8Y8BslwR",
+      "text" : "LNCEqkGa3Y8DcWRKos",
+      "author" : "0hjH2mlLZWfjhDkEbB5",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/idfacere",
+  "link" : "https://www.example.org/doloresvoluptates",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "pzGP7WKwf5QD2Y",
+    "mainTitle" : "g2Fo8VK6HN",
     "alternativeTitles" : {
-      "is" : "gS6FQibnfgdbbLE"
+      "ru" : "JT91UVLaojHrbZ3Z1"
     },
-    "language" : "http://lexvo.org/id/iso639-3/pol",
+    "language" : "http://lexvo.org/id/iso639-3/zho",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "xBtky40RlXKyGA",
-      "month" : "7RVBh18RoMpB8",
-      "day" : "XQdMP4MNQMO"
+      "year" : "9R5R8dXfHI",
+      "month" : "0MEIG9LZhUee",
+      "day" : "w2IFXYKY2hmIHmBwaIa"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/s9FJiRxFkG",
-        "name" : "zpzisGMsOh6zz",
-        "nameType" : "Organizational",
-        "orcId" : "l5OWgj7vfpCbN",
-        "arpId" : "CaaHi62HSl4M9m"
+        "id" : "https://www.example.com/T7NKjWKZkz",
+        "name" : "8krkh8hF6QwqGPhlv",
+        "nameType" : "Personal",
+        "orcId" : "iTYoSQFp1C",
+        "arpId" : "QjIqRF8qsYYRUw6"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/15lMJ23PcxftYELQ3",
+        "id" : "https://www.example.com/CXG0RGMkkva",
         "labels" : {
-          "zk8338A7dTuR" : "miO6CrD7E3"
+          "ndVyOkoa2Isx5" : "3ypjrgO2HFn"
         }
       } ],
-      "role" : "Creator",
-      "sequence" : 9,
+      "role" : "Journalist",
+      "sequence" : 0,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/LjLMKA8ZYAKQ",
-        "name" : "JApeR9opcVkyV7183",
+        "id" : "https://www.example.com/93C7xdbzWKlKu58C",
+        "name" : "Qfl6KtrxwSMi86vz",
         "nameType" : "Personal",
-        "orcId" : "anC6akr6e84NziIT",
-        "arpId" : "F47bD4OVNCkPz7R9hOm"
+        "orcId" : "fPRxphz7gKM",
+        "arpId" : "8hF3QwuiugY4"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/uRNYPlQbOfyVhKOS3v",
+        "id" : "https://www.example.com/RnXDXHBuiZeP",
         "labels" : {
-          "sslTCFIZdKTa" : "nmyAMNgksVXtQe3nT"
+          "i94MoaRv4bzf" : "V7PzQO6O9nYpfr"
         }
       } ],
-      "role" : "ProgrammeParticipant",
+      "role" : "Sponsor",
       "sequence" : 9,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "VFJckl8jAcRW7Q801",
-    "tags" : [ "nCIXTsw0gHZVcCdQr" ],
-    "description" : "TFLpELKikZznfd",
+    "npiSubjectHeading" : "Woh9DPbkoiUM5V4R",
+    "tags" : [ "HliQ7YJ1tPBW2KB" ],
+    "description" : "WFY3QYsQ6fDv",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Chapter",
-        "partOf" : "https://www.example.com/D8mqm06WLELdPKUqghk"
+        "partOf" : "https://www.example.com/iDyxlBHlj3Pb29dbe"
       },
-      "doi" : "https://www.example.com/ttUcCM7VgSokbN7aXNO",
+      "doi" : "https://www.example.com/qk162AJzWbH",
       "publicationInstance" : {
         "type" : "ChapterArticle",
-        "contentType" : "Textbook Chapter",
+        "contentType" : "Academic Chapter",
         "peerReviewed" : false,
         "originalResearch" : true,
         "pages" : {
           "type" : "Range",
-          "begin" : "FaaWROqjQ0xBHvSCCGG",
-          "end" : "XZmoeksNiGg"
+          "begin" : "TMJZByjqwqYzuTjdI",
+          "end" : "heC0GmAtfVi"
         }
       }
     },
-    "metadataSource" : "https://www.example.com/OIHowj9BMBTS08uI7T",
-    "abstract" : "dqLFpl2eoEaPg"
+    "metadataSource" : "https://www.example.com/cjynRqHJtLyalvct",
+    "abstract" : "YF8S7YYE989r2RtrkV"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "4mf5JFOb93sw0lHZqz",
-      "mimeType" : "fGzU3fTN9OPa5Hfmbr9",
-      "size" : 1458613597,
+      "name" : "5AK6eZxgHL",
+      "mimeType" : "tRowdG603tm3U",
+      "size" : 568577536,
       "license" : {
         "type" : "License",
-        "identifier" : "OH3MYiujWCJXoHzd",
+        "identifier" : "3UoLwgiyTz",
         "labels" : {
-          "Y8iaonzm8mT1vE" : "FbA7NvRIAf9Ikoh8"
+          "o4Y8ZyGjtK" : "rlZZuHSMkcZF4gw"
         },
-        "link" : "https://www.example.com/FFfqHh8y6KeoJ1"
+        "link" : "https://www.example.com/qTamfqfYosdVjzYR"
       },
-      "administrativeAgreement" : false,
+      "administrativeAgreement" : true,
       "publisherAuthority" : false,
       "embargoDate" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/reiciendisqui",
-    "name" : "JThOWS8YOf",
+    "id" : "https://www.example.org/oditrepellat",
+    "name" : "byAuFCrwJVUHA9zfm",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "47AJOWSnjcutC",
-      "id" : "9WjdJFKmujBUd"
+      "source" : "DKWu2XrhwAKZEL50j",
+      "id" : "UPXTRD9LKCWWAA"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
-      "approvalStatus" : "REJECTION",
-      "applicationCode" : "4EcfWmKZlwm"
+      "approvedBy" : "NMA",
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "vptqtSvhD0JQCn0cjS"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -155,6 +155,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/debitisperferendis" ],
+  "subjects" : [ "https://www.example.org/sequieligendi" ],
   "modelVersion" : "0.14.24"
 }
