+++ documentation/ConferenceLecture.json
@@ -2,103 +2,103 @@
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "NEW",
-  "owner" : "smX535KO7q",
+  "owner" : "ATv7l24TwDKB4Z",
   "resourceOwner" : {
-    "owner" : "smX535KO7q",
-    "ownerAffiliation" : "https://www.example.org/voluptatibusnihil"
+    "owner" : "ATv7l24TwDKB4Z",
+    "ownerAffiliation" : "https://www.example.org/providentmagni"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/inventorepariatur",
+    "id" : "https://www.example.org/molestiassimilique",
     "labels" : {
-      "A8YCLIKsKBPzN" : "GZgcl4KdYwOjM2bcnbA"
+      "obaO59SAGV4A2eg2Pw1" : "lIbmQThGsz8km"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/providenteos",
-  "doi" : "https://doi.org/10.1234/vero",
+  "handle" : "https://www.example.org/suntcumque",
+  "doi" : "https://doi.org/10.1234/dicta",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "APPROVED",
+    "status" : "REJECTED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "sSOhrGNKcVeN",
-      "author" : "Z0pNEgG3UPfmPN01o8",
+      "text" : "bxVsS9ltD7kPwXiJ4v",
+      "author" : "qQEjDNCvVWQlKfsoqLP",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/enimaliquam",
+  "link" : "https://www.example.org/ipsammolestiae",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "6HnpffVp6QhJUl",
+    "mainTitle" : "muVXXH7PGou",
     "alternativeTitles" : {
-      "is" : "Bq38qLA6e6PhcRRyc"
+      "de" : "y9XUsLkgi8kwYw1FjKH"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "z0WWPGCWIEJR",
-      "month" : "0rpqYLhf6s5SgIxlK0",
-      "day" : "u7YnuF0DcqqCNsluxI"
+      "year" : "U0v3Gjf4fgdtRi1e",
+      "month" : "HsdZCMUqzo268",
+      "day" : "S1rKY9ruYvXX95U"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/X0SOwZ2QtXaKcN",
-        "name" : "9IS6m8pR9Q8vd",
-        "nameType" : "Personal",
-        "orcId" : "9jJjVVMBZKYaKlsq",
-        "arpId" : "zWN7lqxI1sx5gVUr"
+        "id" : "https://www.example.com/UDM19Y9nQEbGrHt",
+        "name" : "MyE1ZK8498u1Dmexi",
+        "nameType" : "Organizational",
+        "orcId" : "szbZSioOleFTJEj0ZYC",
+        "arpId" : "iOa9Qs6BkMP9ZG50hs"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/7B77ml9S9aP",
+        "id" : "https://www.example.com/5MH2HX8bDc4cA6U",
         "labels" : {
-          "berP60pxpAhqdaGGi5" : "qi1uzzISq35g5"
+          "paTFAaXu5BXtwbp" : "oyRHACRWfpGO52LD8"
         }
       } ],
-      "role" : "RightsHolder",
-      "sequence" : 6,
+      "role" : "Illustrator",
+      "sequence" : 2,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/3LxA31x43Dktzy",
-        "name" : "y00WOEfr5vMvnFTBQ",
+        "id" : "https://www.example.com/8IG5jTi9dOGwLLOQc",
+        "name" : "H7oQ2Y0kzSBWDsFC",
         "nameType" : "Organizational",
-        "orcId" : "gyLgUR5MnwakxwNhjt",
-        "arpId" : "fIen71k2SPMpl"
+        "orcId" : "GysTiz6yN3YBj",
+        "arpId" : "jziZcADAL5KmQXOMT"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/who3DrAcIvAnMKdF",
+        "id" : "https://www.example.com/klXgnLC71I",
         "labels" : {
-          "ZnTbFRpCMKP" : "tvZHpVFnnCr"
+          "z6ANO8MqQsYtWW" : "9Kdr5VW2yPmPZ"
         }
       } ],
-      "role" : "RelatedPerson",
-      "sequence" : 8,
+      "role" : "DataCollector",
+      "sequence" : 2,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "kYxSAlkN3hzWhM3dy",
-    "tags" : [ "Axi9XVl3NOwor" ],
-    "description" : "Gcccjb0gbP1Znb",
+    "npiSubjectHeading" : "4TdZ8ZsBQ61nFtThl",
+    "tags" : [ "XNHBZ4dzcHsxrcixn" ],
+    "description" : "2Qeydbs7JX928NDZJj",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Event",
-        "label" : "PrAi80NRKo",
+        "label" : "oaxgWvYQxC8p47u",
         "place" : {
           "type" : "UnconfirmedPlace",
-          "label" : "P1bV4eyCgg",
-          "country" : "iOzAxtmDizK"
+          "label" : "dAdwURlhpSxXSlEWl",
+          "country" : "Nw0rDqgPrfy0wB"
         },
         "time" : {
           "type" : "Instant",
@@ -106,14 +106,14 @@
         },
         "agent" : {
           "type" : "Organization",
-          "id" : "https://www.example.com/q9smpXcD7M9zge",
+          "id" : "https://www.example.com/9RbRuMKtlyxmOx8Wg",
           "labels" : {
-            "0Vj3qZaaUtIoS" : "sCVheX5dFROL3nyJzjy"
+            "qxLglgVFxVb5W" : "foDLFEm7bgsjDOL"
           }
         },
-        "product" : "https://www.example.com/u1bL9kQ8XmtEhxP"
+        "product" : "https://www.example.com/d3UmxvbjqsMYpkIcX"
       },
-      "doi" : "https://www.example.com/5vFgTdSboniES",
+      "doi" : "https://www.example.com/f256kijfA0GPyP6l",
       "publicationInstance" : {
         "type" : "ConferenceLecture",
         "peerReviewed" : false,
@@ -122,45 +122,45 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/YEj3kHYFp89u47H",
-    "abstract" : "8v9wvmqLZoD7Y"
+    "metadataSource" : "https://www.example.com/Ktox5RT4MbpXPS",
+    "abstract" : "2EF3jCJ45lxWnh"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "az7ccyNZ4x0cczc",
-      "mimeType" : "c9cKDmsRga0Z7itii9",
-      "size" : 1556300317,
+      "name" : "Jv7PHx46fUqs1BlP3vl",
+      "mimeType" : "X0AqlpbYmhept2tUt",
+      "size" : 1551848865,
       "license" : {
         "type" : "License",
-        "identifier" : "pfSP8ojqoXEtdk26ksT",
+        "identifier" : "Dt4jegG95z",
         "labels" : {
-          "X2YXJiTGsWhV1" : "6iykO935iEA"
+          "vWSdIARhoHCge2LB" : "yereXz1ugUsd"
         },
-        "link" : "https://www.example.com/D56oCsCKw6Ueit"
+        "link" : "https://www.example.com/3gMuDdUqWAU"
       },
-      "administrativeAgreement" : true,
+      "administrativeAgreement" : false,
       "publisherAuthority" : false,
       "embargoDate" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/idnisi",
-    "name" : "0dkugLaJyZ3",
+    "id" : "https://www.example.org/facilisvelit",
+    "name" : "FRldz9beZZlz6zz",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "XDfRHhy95hKne0",
-      "id" : "aFEAMk1UuacA"
+      "source" : "OYbD6FNhBjHVJHTXNTc",
+      "id" : "m5AZtuoAKR6ihVpFzX"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "vzhRinz6eQGj"
+      "approvedBy" : "REK",
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "yQfe518fxv"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -168,6 +168,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/eiusquia" ],
+  "subjects" : [ "https://www.example.org/ipsamqui" ],
   "modelVersion" : "0.14.24"
 }
