+++ documentation/ConferencePoster.json
@@ -1,119 +1,120 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
-  "owner" : "pF9sQIKMjO",
+  "status" : "DRAFT_FOR_DELETION",
+  "owner" : "ob7dzWxzsspFYXGfPD",
   "resourceOwner" : {
-    "owner" : "pF9sQIKMjO",
-    "ownerAffiliation" : "https://www.example.org/voluptasrerum"
+    "owner" : "ob7dzWxzsspFYXGfPD",
+    "ownerAffiliation" : "https://www.example.org/quodoccaecati"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/rationenisi",
+    "id" : "https://www.example.org/namhic",
     "labels" : {
-      "hqQfYxsFcsBK" : "seusxG9a7iLdLbJ"
+      "1VfLkofgndrUXBn" : "uIcFUXt2YWOmdGuUaq"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/natusillum",
-  "doi" : "https://doi.org/10.1234/in",
+  "handle" : "https://www.example.org/placeatconsequatur",
+  "doi" : "https://doi.org/10.1234/quo",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REJECTED",
+    "status" : "REQUESTED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "X0GrZAOTFXQYL",
-      "author" : "1AS4wSUinccsc1Qc",
+      "text" : "1Nhs5CabBB7o62",
+      "author" : "CYaBla4GVsJ",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/enimillum",
+  "link" : "https://www.example.org/estfugit",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "cSRGYdrVBHRu",
+    "mainTitle" : "sNQAUM1uI4Hdbx",
     "alternativeTitles" : {
-      "nl" : "rY3PS6UV0y"
+      "it" : "bNg7Q2B3PqjfajM3TZ"
     },
-    "language" : "http://lexvo.org/id/iso639-3/sme",
+    "language" : "http://lexvo.org/id/iso639-3/pol",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "n8Upq7E47QZvd",
-      "month" : "P0QLgQ8vdTaBr3cCCYL",
-      "day" : "bDFaM3l6QDhrGoUH"
+      "year" : "Sa6ypx2WMUYmquFbo",
+      "month" : "yWJzBFVwq8FRe",
+      "day" : "pYNPSTKoBTMLHKilfBc"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/8yAceFpVLjYaKb0oXfx",
-        "name" : "iVhYaApxTEFpI",
+        "id" : "https://www.example.com/DGvC4bdvmhCvqGQgH",
+        "name" : "6U3ym7sS2Cqp2SOI8V",
         "nameType" : "Organizational",
-        "orcId" : "aW1KI9K0NXls",
-        "arpId" : "StFAZUGkCZxM4G"
+        "orcId" : "5Hm3UZiX5Uu",
+        "arpId" : "2KLhfrQzSJuUZdj8"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/yd2pw5pofuEqaIu4r",
+        "id" : "https://www.example.com/aFkPa0AGr49le2YL5Hl",
         "labels" : {
-          "2wNGJV5kQsf6edS5P9" : "S13TkCYjl4nnEv"
+          "qhdbVULZRyY" : "q1QlYsTXJYx2"
         }
       } ],
-      "role" : "Researcher",
-      "sequence" : 7,
+      "role" : "AcademicCoordinator",
+      "sequence" : 2,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/8AEjCCg0UeZ7sA4CffA",
-        "name" : "USTKk6NIKxVfoRj",
+        "id" : "https://www.example.com/iAdf2oQZztpqTq26lH",
+        "name" : "blcFuSgTDlSc9SP",
         "nameType" : "Organizational",
-        "orcId" : "RuvY5ceMnGzMcvHfxqm",
-        "arpId" : "8BDmI9g0Re"
+        "orcId" : "TyWHWgSU9Byq0aFNRl0",
+        "arpId" : "XZVCloMuMd3"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/YDmGeZXN4K",
+        "id" : "https://www.example.com/9KdOXjCDr1Tz8WzL4lS",
         "labels" : {
-          "GCZRuTJ1srHnQl8u4O6" : "heTAGRdbo8MyJzsuDQr"
+          "NGbx5WjGwAw" : "aidS5N3b5s"
         }
       } ],
-      "role" : "RightsHolder",
-      "sequence" : 1,
+      "role" : "Funder",
+      "sequence" : 0,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "MajikFAtDzuAtbyXXD",
-    "tags" : [ "NqPwrnttOA" ],
-    "description" : "JAX1XmpoGHpaFvIMq",
+    "npiSubjectHeading" : "y0SGtwRNHCPF8anJ7",
+    "tags" : [ "OE42CLGMZQ" ],
+    "description" : "dQ4ioYXKXqs6VAlE",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Event",
-        "label" : "osehwQINRpBmA8IXgdm",
+        "label" : "M0SSZh9q6bPN",
         "place" : {
           "type" : "UnconfirmedPlace",
-          "label" : "F5amr7tPJ5zKe",
-          "country" : "0w0G9ruvsJF"
+          "label" : "6Bh4FZ6mr2dLt3mw",
+          "country" : "vVRBcKjstgWC4sCO"
         },
         "time" : {
-          "type" : "Instant",
-          "value" : "2020-09-23T09:51:23.044996Z"
+          "type" : "Period",
+          "from" : "2020-09-23T09:51:23.044996Z",
+          "to" : "2020-09-23T09:51:23.044996Z"
         },
         "agent" : {
           "type" : "Organization",
-          "id" : "https://www.example.com/6PHGsQyUODLzlx",
+          "id" : "https://www.example.com/ussyOAnwOd6",
           "labels" : {
-            "YRsTwlc6UHPp3N9" : "fsnGOLhr1qxe"
+            "VDgxZjCXQeTG" : "mLkIWl1PihgvmnL5P8"
           }
         },
-        "product" : "https://www.example.com/iGdJL6xKSaZ210H"
+        "product" : "https://www.example.com/CCLPXZT0wjq74"
       },
-      "doi" : "https://www.example.com/Wm4sMZwM5y5C",
+      "doi" : "https://www.example.com/Of8FGHxD4b1r",
       "publicationInstance" : {
         "type" : "ConferencePoster",
         "peerReviewed" : false,
@@ -122,24 +123,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/wsw9GGh8D4M",
-    "abstract" : "zLpu3YJW9Zrbft1mS"
+    "metadataSource" : "https://www.example.com/0XroT7oHOmhlx",
+    "abstract" : "S837y61KSzu"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "qYKarESa7zJmrRKlf",
-      "mimeType" : "rjGoExWQPvHKdA",
-      "size" : 1256119483,
+      "name" : "ykS5GV9vgupk",
+      "mimeType" : "PYRvlA71BBU2",
+      "size" : 1143301598,
       "license" : {
         "type" : "License",
-        "identifier" : "cMZjx6GRqL9zdEL8YWT",
+        "identifier" : "74wH5LqaV9A",
         "labels" : {
-          "Cq4J8sdrpaDeh" : "ISl2oWQp2SMamlF"
+          "FTSOfgxqdBvo7KX" : "VlXrJoql11oPlF"
         },
-        "link" : "https://www.example.com/keiGHcODvNHxmp4PQ"
+        "link" : "https://www.example.com/whujXU15cj"
       },
       "administrativeAgreement" : false,
       "publisherAuthority" : false,
@@ -148,19 +149,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/voluptaseos",
-    "name" : "uWUMS7jaVgHRaII",
+    "id" : "https://www.example.org/illoconsequatur",
+    "name" : "WRQsIgj1sNF9XG0no",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "1UzL1OU5w00",
-      "id" : "1wM4dtJ62aJ3"
+      "source" : "Ix6FqCbNlsLgQxenSG",
+      "id" : "dZAk7jI5gkXCmuPgt"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "16XsLFVZrNkUU6T"
+      "approvedBy" : "NMA",
+      "approvalStatus" : "APPROVED",
+      "applicationCode" : "JrBKSr2dQvW3mrY"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -168,6 +169,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/esseet" ],
+  "subjects" : [ "https://www.example.org/eosexcepturi" ],
   "modelVersion" : "0.14.24"
 }
