+++ documentation/FeatureArticle.json
@@ -1,154 +1,154 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
-  "owner" : "hNZzKJUS5ZPk3e",
+  "status" : "NEW",
+  "owner" : "1QMXSgDWK4IWKa",
   "resourceOwner" : {
-    "owner" : "hNZzKJUS5ZPk3e",
-    "ownerAffiliation" : "https://www.example.org/quaeratcupiditate"
+    "owner" : "1QMXSgDWK4IWKa",
+    "ownerAffiliation" : "https://www.example.org/sedqui"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/saepereiciendis",
+    "id" : "https://www.example.org/dolorumfugiat",
     "labels" : {
-      "XgB7kp2BVV" : "7HgGcueQwlo"
+      "abmPaONuTLMnopUSXj" : "clZrLcsG7hUbM3W"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/quaeratvelit",
-  "doi" : "https://doi.org/10.1234/voluptatibus",
+  "handle" : "https://www.example.org/doloribusitaque",
+  "doi" : "https://doi.org/10.1234/ad",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REQUESTED",
+    "status" : "REJECTED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "8nirjCgLh34eSVGZRc",
-      "author" : "1mUqP9x5yFSEuwbjTu",
+      "text" : "epxpT0uuri1lTVSaTzy",
+      "author" : "hUV3nnIy2YWFnxfj",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/eaquequia",
+  "link" : "https://www.example.org/delenitidicta",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "GkiBRy2Fi102",
+    "mainTitle" : "e04sZEfkiYhBSKSTZ",
     "alternativeTitles" : {
-      "zh" : "wjgFAdgAYvFFaY"
+      "ru" : "Tl35L17MYovTf7gi6"
     },
     "language" : "http://lexvo.org/id/iso639-3/rus",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "o2JRR6mOkV36KXhQy",
-      "month" : "Soi93BMMtnHyW5nEY",
-      "day" : "9WA680ZqvZMPhQ7d"
+      "year" : "Cw77YnIIxCggaFxD",
+      "month" : "2QWgKjGHgfjKf1DX86A",
+      "day" : "fZqJWMY5eOS"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/AlBkes4pAtTe7S",
-        "name" : "4IrT9LjlM9xg",
+        "id" : "https://www.example.com/yvFpuQhV8WFUKN",
+        "name" : "flrd3k1Vp5WR4",
         "nameType" : "Personal",
-        "orcId" : "EyQxGj59TQf",
-        "arpId" : "NNldNPkAt4"
+        "orcId" : "tIvLYISOYx",
+        "arpId" : "V6WtvFC48W0INySrSKe"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/58cIJwREJ4lIta",
+        "id" : "https://www.example.com/Bgo0nxakoUe98",
         "labels" : {
-          "ExBhfX38mcMt" : "VTkVKStuOE8x"
+          "Se0ku3Q2eKgzBC" : "hTtnG0AIn6B632"
         }
       } ],
-      "role" : "ProjectMember",
+      "role" : "Sponsor",
       "sequence" : 7,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/wkYBjinyAWu",
-        "name" : "pZmuZGzZ6SRhwCsSFme",
+        "id" : "https://www.example.com/j2GEuzklnuGv",
+        "name" : "Dp7WfAiBvgSQmeldNA",
         "nameType" : "Personal",
-        "orcId" : "ppaiS8JhPCw1VE",
-        "arpId" : "Dp1mM39dTikRtFVEf"
+        "orcId" : "GvyLeFpHBb",
+        "arpId" : "kqGErHx3QM"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/FqbF3PJjvUegS6pgMp7",
+        "id" : "https://www.example.com/W34c0clE1wHebY46F",
         "labels" : {
-          "MBVLGxGx4Rd99XOqs" : "JXqVpkAnJcT1jEr9hw2"
+          "wcByoUuLuPcY" : "WsOP2lZDuNirJTxq"
         }
       } ],
       "role" : "Editor",
-      "sequence" : 1,
+      "sequence" : 9,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "GyryZPt9CeUHMQ",
-    "tags" : [ "c19qJ8CuZjpralKK" ],
-    "description" : "fW72Uj1zgZ",
+    "npiSubjectHeading" : "WNl2tG4ZYZtIOo",
+    "tags" : [ "voDiPf0d9MJ" ],
+    "description" : "rLvgc7Fp708fGn7",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/PcgbOjOneGr3XWPHO"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/9sGssNII5pYI8qRmwu"
       },
-      "doi" : "https://www.example.com/mrQdyurc8z1tW",
+      "doi" : "https://www.example.com/rZkSow8aZEO6",
       "publicationInstance" : {
         "type" : "FeatureArticle",
-        "volume" : "p1MiHj03K8",
-        "issue" : "H3X5jPz9VM6Vbe",
-        "articleNumber" : "4qaNxocfy9dS",
+        "volume" : "QpU8toDB98Fn5Re0",
+        "issue" : "CA6Y8bB8y3hlxl",
+        "articleNumber" : "DmHBWM89ey2hsq1n",
         "pages" : {
           "type" : "Range",
-          "begin" : "3LYyPXrPP2KrM",
-          "end" : "GDFkkEvKtO"
+          "begin" : "ixAghdiFHkPa7WKD1f",
+          "end" : "H0YAoOYYrVSMVU"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/j5IqcKFMcB",
-    "abstract" : "AkMDWAevRXeq"
+    "metadataSource" : "https://www.example.com/UKdmPvjRq2EQQ7",
+    "abstract" : "dLU1sOkHvfmn5M"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "2LujiM5gua",
-      "mimeType" : "xF9C0ZCeIQj2V",
-      "size" : 485333305,
+      "name" : "xQW7WR8qST1bU3sqk86",
+      "mimeType" : "DnEDiWu5iBe1d6",
+      "size" : 1915281206,
       "license" : {
         "type" : "License",
-        "identifier" : "cnseLZaK5myZ",
+        "identifier" : "vyUliArRSZxjk2sYoL",
         "labels" : {
-          "7FPSRA4MA4NF1mVdZCI" : "VQdJpVU4YL6Sb9gWnxp"
+          "daXp0KHPT5ZkHp" : "19Y94XyRQVCUF0cY"
         },
-        "link" : "https://www.example.com/d7a0Wt0YFn6ZFqf"
+        "link" : "https://www.example.com/tDBHagmo1fpBNJrAK"
       },
-      "administrativeAgreement" : true,
+      "administrativeAgreement" : false,
       "publisherAuthority" : false,
       "embargoDate" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/consequatursaepe",
-    "name" : "7iLUJd9Z8qO",
+    "id" : "https://www.example.org/beataereprehenderit",
+    "name" : "MzBB5pIQIUI97XSx",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "PTI9YgDNWyme",
-      "id" : "By6l5ZsxQ3bXva75"
+      "source" : "1NY1Eb8pNUyT3Gw",
+      "id" : "prW1SSZ6YfKnI510Ce6"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
+      "approvedBy" : "REK",
       "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "gFiUY3UlFF77kSgd"
+      "applicationCode" : "gj7A8WLGWzZoErJGHE"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/estrerum" ],
+  "subjects" : [ "https://www.example.org/utsed" ],
   "modelVersion" : "0.14.24"
 }
