+++ documentation/JournalArticle.json
@@ -1,25 +1,25 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "NEW",
-  "owner" : "7YTPXuUOQtzo",
+  "status" : "DRAFT",
+  "owner" : "b4EnQ5Rw7dVY68NSt1m",
   "resourceOwner" : {
-    "owner" : "7YTPXuUOQtzo",
-    "ownerAffiliation" : "https://www.example.org/deseruntporro"
+    "owner" : "b4EnQ5Rw7dVY68NSt1m",
+    "ownerAffiliation" : "https://www.example.org/utdeleniti"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/officiaut",
+    "id" : "https://www.example.org/nobisquod",
     "labels" : {
-      "R17BDocDbm" : "igWnbC9nLSZ8"
+      "TcLAmOMWRRfycRUMmi" : "Zjj778x01S7"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/rerumnihil",
-  "doi" : "https://doi.org/10.1234/necessitatibus",
+  "handle" : "https://www.example.org/temporenisi",
+  "doi" : "https://doi.org/10.1234/quo",
   "doiRequest" : {
     "type" : "DoiRequest",
     "status" : "APPROVED",
@@ -27,109 +27,109 @@
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "uZo4zkfs7v",
-      "author" : "QCg6xQLsT1NKP",
+      "text" : "G5nl3XSn6hM73djS",
+      "author" : "HHZibtc8uhav1omNsmb",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/erroret",
+  "link" : "https://www.example.org/harumut",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "7FljzBjuTaWqsG5q",
+    "mainTitle" : "WSTAptmUVyImG238",
     "alternativeTitles" : {
-      "de" : "xgOeMvBQZUz"
+      "da" : "DllyKkXtbnih"
     },
-    "language" : "http://lexvo.org/id/iso639-3/nor",
+    "language" : "http://lexvo.org/id/iso639-3/nob",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "cje9qxA5fFKMROApx6b",
-      "month" : "AKD9NDnNfjyD",
-      "day" : "1XaB9LFl8pUw31"
+      "year" : "4mksLlvzfNkR4MS",
+      "month" : "tKZGR7yxDcFXThRwcMM",
+      "day" : "uRhS3kgtf1xx0J"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Xdaab3IMUJNFut3",
-        "name" : "H9Ryp0We9mSQx2Hm",
-        "nameType" : "Personal",
-        "orcId" : "BHpZNy91YY",
-        "arpId" : "7fPFLoR99wRP"
+        "id" : "https://www.example.com/Sk0aTXtl8TKXhqAYH",
+        "name" : "cQPl1RxVqQ2",
+        "nameType" : "Organizational",
+        "orcId" : "dlTIK5xooidwK0BlSXD",
+        "arpId" : "rao8SoSTwJyXXqtN1"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/FHUcnGsTSDb",
+        "id" : "https://www.example.com/R7i17pP6S7uA",
         "labels" : {
-          "HOYHRx7PpnR4S" : "ya54jb9mgJpF"
+          "A81QyKeA1H1QsJir" : "umr7FUiD3Gu70AQyG"
         }
       } ],
-      "role" : "DataCurator",
-      "sequence" : 1,
+      "role" : "Supervisor",
+      "sequence" : 3,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/88KNLz4BJV",
-        "name" : "YfR28WB4tjhvM",
-        "nameType" : "Personal",
-        "orcId" : "IVmrGM1iDF",
-        "arpId" : "wzQr2M6oJww"
+        "id" : "https://www.example.com/KKg5oaEYuQYOH",
+        "name" : "jgwD8X9cxFk8vg4nx",
+        "nameType" : "Organizational",
+        "orcId" : "at4fPyImtIvMRe",
+        "arpId" : "3KkaZTr95l"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/UkLH6WMFeIUiisbCeY",
+        "id" : "https://www.example.com/isohyPYQOn69IX77",
         "labels" : {
-          "460anXTlBkJyvqQRTOn" : "N4uKjIgjYRJKrXAaia"
+          "JKvmHYuLG9vcZF" : "IlihMIYb12"
         }
       } ],
-      "role" : "HostingInstitution",
-      "sequence" : 7,
+      "role" : "ResearchGroup",
+      "sequence" : 3,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "li21bjAQsnagC6lhNgb",
-    "tags" : [ "7ZHyfF0PH6AbG7T" ],
-    "description" : "6w5mkwQs8ic2n",
+    "npiSubjectHeading" : "mJ1a6jKIs4aU",
+    "tags" : [ "O7Nk0m9QdwF" ],
+    "description" : "EMT3iQeLyX7P9PWAJ",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/YNSWgugYLf3U6ZQ89"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/JpjorHgzmpZla5"
       },
-      "doi" : "https://www.example.com/bRiv5hVIGMOVP",
+      "doi" : "https://www.example.com/5UpWUWP9QiZWYP",
       "publicationInstance" : {
         "type" : "JournalArticle",
-        "volume" : "ofvgTQBUv4Z",
-        "issue" : "aSekRiWe66eVXg1M",
-        "articleNumber" : "zdmXWPbRiA",
-        "contentType" : "Professional article",
+        "volume" : "HJTfG5pn6Zr4n",
+        "issue" : "ik2miTdvXP7enh",
+        "articleNumber" : "BIR7myopGThUspdtBN7",
+        "contentType" : "Case report",
         "peerReviewed" : false,
         "originalResearch" : false,
         "pages" : {
           "type" : "Range",
-          "begin" : "d8lLlbT6YyBkzm2Oqq",
-          "end" : "6nnXwxcY98c"
+          "begin" : "UuqZHBhKb37",
+          "end" : "D7o5Cd1pd4AsXlJ9N4"
         }
       }
     },
-    "metadataSource" : "https://www.example.com/gqdPssOAyUs",
-    "abstract" : "X2DSD9KgFrOMWFGHE5"
+    "metadataSource" : "https://www.example.com/fzd3EOlklaedT",
+    "abstract" : "pu3EVMKUMZq"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "kM0m1OHk16",
-      "mimeType" : "17e8GfEvx9ltJjWm",
-      "size" : 1999564737,
+      "name" : "9cApVzzWHQXuxEN",
+      "mimeType" : "XFY9H3PEZAT",
+      "size" : 1460867371,
       "license" : {
         "type" : "License",
-        "identifier" : "43qbdgg7ACh",
+        "identifier" : "Lln9GwLJZ6BclVZ",
         "labels" : {
-          "XAN0at7BJkqr5U" : "KlMcMtqtx8"
+          "v0yJ8KCsZdf" : "kHBvFFDKaX3snQNI"
         },
-        "link" : "https://www.example.com/PS1CKBo6oK2JSeN"
+        "link" : "https://www.example.com/0Xbuu3YTPYetyqWj"
       },
       "administrativeAgreement" : false,
       "publisherAuthority" : false,
@@ -138,19 +138,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/liberonemo",
-    "name" : "WpbCmNOZGljDFXuU",
+    "id" : "https://www.example.org/doloremtotam",
+    "name" : "ZWh9EmKNvRfP5",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "Rw4rR4Vzkd",
-      "id" : "uBJLSAw3DVcAf5NYPo"
+      "source" : "l5wgdvS59nJNRMpEGg",
+      "id" : "VHceg0KF4rKQjWUe"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "APPROVED",
-      "applicationCode" : "mVPscbdsYblKwAL6M"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "APPLIED",
+      "applicationCode" : "lU8tlt276cgLNh4"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -158,6 +158,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/rerumex" ],
+  "subjects" : [ "https://www.example.org/delectusreprehenderit" ],
   "modelVersion" : "0.14.24"
 }
