+++ documentation/JournalCorrigendum.json
@@ -1,134 +1,134 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
-  "owner" : "nwe2340ljAWdjSSF",
+  "status" : "NEW",
+  "owner" : "DpDvm4MjMtmW57clGMI",
   "resourceOwner" : {
-    "owner" : "nwe2340ljAWdjSSF",
-    "ownerAffiliation" : "https://www.example.org/estet"
+    "owner" : "DpDvm4MjMtmW57clGMI",
+    "ownerAffiliation" : "https://www.example.org/hicaut"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/dolorut",
+    "id" : "https://www.example.org/aliasnumquam",
     "labels" : {
-      "4Fitdt4YfeD" : "6yStVcI5kYVDK2wnCyf"
+      "sJhYIbU1AvK4G9" : "CK6l2uAfy6y9kCw"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/autexpedita",
-  "doi" : "https://doi.org/10.1234/omnis",
+  "handle" : "https://www.example.org/quasveritatis",
+  "doi" : "https://doi.org/10.1234/nesciunt",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REJECTED",
+    "status" : "REQUESTED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "UEHEKXUP94Gw7R61m",
-      "author" : "BXaGMa2Wmp",
+      "text" : "tTNszR2Yxbq6",
+      "author" : "h7AiFlzV2I",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/eligendiconsequatur",
+  "link" : "https://www.example.org/quisnon",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "mC8iRbdqhvzMebw5",
+    "mainTitle" : "BjVcsTgCSSnnbntkHUA",
     "alternativeTitles" : {
-      "pt" : "WGrlts6D2jOuSDtI"
+      "bg" : "eivMrK8Xxs"
     },
-    "language" : "http://lexvo.org/id/iso639-3/afr",
+    "language" : "http://lexvo.org/id/iso639-3/pol",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "tBbAeSs3biz9uA",
-      "month" : "qPXgGcU7ZYOA0Df3",
-      "day" : "sQ6UXdl9bvG"
+      "year" : "JnrPk8PkQCpdeGLI",
+      "month" : "trVjUoOhrsNlb1dR",
+      "day" : "EDQ2mpDUHP5DR1Q3"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/DkfBI5DKG3wVHEVOg6g",
-        "name" : "xU03OY04pFmOTv",
+        "id" : "https://www.example.com/mKECwjmxJ5Mkse",
+        "name" : "Alr5qjmX9uWcgf",
         "nameType" : "Organizational",
-        "orcId" : "FoPcg2NZ7gP0Q96kGH8",
-        "arpId" : "aMRwe67NKKCvq"
+        "orcId" : "WOAW8ctoChCg",
+        "arpId" : "ibHpla8YdlByi"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/uV2RQpv2OR39M",
+        "id" : "https://www.example.com/XrpMjIQFJfY",
         "labels" : {
-          "ezbpD0ZlhtC" : "RSZzXMAMlDQw19b"
+          "yMnmLz7USXowtzw2XbG" : "e9e4xH9mXea2uOmbm"
         }
       } ],
-      "role" : "Other",
-      "sequence" : 8,
+      "role" : "RegistrationAgency",
+      "sequence" : 1,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/RRXHtfW1DuRYC",
-        "name" : "pIjHBFGxBe",
-        "nameType" : "Personal",
-        "orcId" : "nan0S77UholXZ",
-        "arpId" : "pmuWig8EZ2n0ImHeF"
+        "id" : "https://www.example.com/GmCbuxYAZUmfFr0Zx",
+        "name" : "GleEuYLqoXd7WKo",
+        "nameType" : "Organizational",
+        "orcId" : "6lmRjGfDnm7qq",
+        "arpId" : "5R1hK6KYPWZGDY"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/lYmZ1UiZYgzX6X",
+        "id" : "https://www.example.com/HcVqglTML3SwsRjkpf",
         "labels" : {
-          "1YwHZatD7m" : "ISsD6rjS5KSQ"
+          "w7JGx10rbOe8ZvfaHg" : "NPpxaXvNyXuQ"
         }
       } ],
-      "role" : "CuratorOrganizer",
-      "sequence" : 6,
+      "role" : "RegistrationAgency",
+      "sequence" : 2,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "RAyVuccDhQlgY",
-    "tags" : [ "OGXW8OSMs6QXSuTzWO" ],
-    "description" : "gLNtTblIsqBh",
+    "npiSubjectHeading" : "sovNS0TCaQgZ",
+    "tags" : [ "JcffleZqXi0vuKhD" ],
+    "description" : "GBIyJZJbcPNvePWeP",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/2Lv56PmQZ6BqXCgn"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/5hZ4ezDdDrMNDrblMs7"
       },
-      "doi" : "https://www.example.com/pTH85rrp4R8A",
+      "doi" : "https://www.example.com/GT143r67rPaGFvNo7T4",
       "publicationInstance" : {
         "type" : "JournalCorrigendum",
-        "volume" : "61pD80cAjLXsrtoHk",
-        "issue" : "Wa2k8jPo3dNiIb3lf",
-        "articleNumber" : "mlzyrlkuvXzlF",
+        "volume" : "h7dwo2OSC6qGqYA6a",
+        "issue" : "pqu7EGzGyQj3JjrV",
+        "articleNumber" : "NHdE8EDDwzKSBL",
         "pages" : {
           "type" : "Range",
-          "begin" : "ishr6ZKE9EJ",
-          "end" : "P8jnzCxM6lGbId5YV2c"
+          "begin" : "8EmEU4Yz1BW",
+          "end" : "LdkDsHUGna"
         },
-        "corrigendumFor" : "https://www.example.com/eW73nMX0lYpb",
+        "corrigendumFor" : "https://www.example.com/DsubAfLjdgu",
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/vy4Hzn26fdo2vaciF",
-    "abstract" : "mWFfRmKYoKTqqs"
+    "metadataSource" : "https://www.example.com/pQTmJYtwKsTn4h",
+    "abstract" : "9zrCBvb5A0I0C"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "qifW8RbXZC8VHbTZ9Z",
-      "mimeType" : "WKlmRgCt2IDnEWFHgh",
-      "size" : 650481143,
+      "name" : "pwKeKkUIhaomdan",
+      "mimeType" : "KjGvuFnTr9JxQLkG",
+      "size" : 182835341,
       "license" : {
         "type" : "License",
-        "identifier" : "HINdpDENNmbvsy",
+        "identifier" : "61vNwfOKFKLn7dH3Zn",
         "labels" : {
-          "PERkwSYXQGvatQz2y" : "ixJ00zVyCFMauUGGC"
+          "fcF0eGEVTv" : "cDrPLNd1ei4Jhmk8Kb"
         },
-        "link" : "https://www.example.com/psaWMHkbTIAyRUEOA9"
+        "link" : "https://www.example.com/id3i6UQczJ8x8L"
       },
       "administrativeAgreement" : false,
       "publisherAuthority" : false,
@@ -137,19 +137,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/etnihil",
-    "name" : "B9r7qUFfYBkIny",
+    "id" : "https://www.example.org/sapientemolestiae",
+    "name" : "n1LS5pRh5hODAQvV9AC",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "BUrE8uOVauGEh8",
-      "id" : "Rn9Fwxj72V1wh"
+      "source" : "c2rYC7Luq5A",
+      "id" : "tQQxeDkXIB0ecnQ"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "nvs9VMQMFroUH"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "JfjDMv5bW3BqhuuWR4"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -157,6 +157,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/autemvoluptatibus" ],
+  "subjects" : [ "https://www.example.org/cupiditateillo" ],
   "modelVersion" : "0.14.24"
 }
