+++ documentation/JournalInterview.json
@@ -1,25 +1,25 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "NEW",
-  "owner" : "dj23oLnitiwYvnLBV",
+  "status" : "PUBLISHED",
+  "owner" : "SmtHuWYixfC",
   "resourceOwner" : {
-    "owner" : "dj23oLnitiwYvnLBV",
-    "ownerAffiliation" : "https://www.example.org/quaetenetur"
+    "owner" : "SmtHuWYixfC",
+    "ownerAffiliation" : "https://www.example.org/porroprovident"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/nobislaboriosam",
+    "id" : "https://www.example.org/aliquiddeleniti",
     "labels" : {
-      "b1bSbL76LKzrQR6rT" : "i9g6O8y1nCM97WvTaiC"
+      "G0ScMgQznPaDLuhDF" : "qbT5B85zKMDXfh"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/siteligendi",
-  "doi" : "https://doi.org/10.1234/impedit",
+  "handle" : "https://www.example.org/eaqueexplicabo",
+  "doi" : "https://doi.org/10.1234/esse",
   "doiRequest" : {
     "type" : "DoiRequest",
     "status" : "REJECTED",
@@ -27,107 +27,107 @@
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "KUX1iHbelmPkX8V",
-      "author" : "H8MHfN0DxVs1w2dgFG",
+      "text" : "qUdRjt0fwu1szCFuJ",
+      "author" : "I8aKzmOgGIzML",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/adet",
+  "link" : "https://www.example.org/doloremqueaperiam",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "1MMvms0Wxt",
+    "mainTitle" : "etH4ICiqKPdjAZr",
     "alternativeTitles" : {
-      "sv" : "1pekK5rjyJI6GCGp"
+      "ru" : "4rP0rg2oSNr"
     },
-    "language" : "http://lexvo.org/id/iso639-3/mul",
+    "language" : "http://lexvo.org/id/iso639-3/bul",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "AEfRR3HtbNJtYBm6e",
-      "month" : "nJNXrVoJA2",
-      "day" : "tY2W9Rfy468J"
+      "year" : "4GNCQEPklTUw",
+      "month" : "U9CYQQOXt8k",
+      "day" : "yHJnRs5KMhqWd"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/wDxZOylop1desSvE",
-        "name" : "oJOyAR8neZIia",
-        "nameType" : "Personal",
-        "orcId" : "5MGbG6kg6tQUCy3o",
-        "arpId" : "oFbC4srSRTdx4AHNx"
+        "id" : "https://www.example.com/KDfQ9ZPXqHY",
+        "name" : "zB0q3YPMkEj",
+        "nameType" : "Organizational",
+        "orcId" : "4IBbtIbPYFJ",
+        "arpId" : "DlfEFPSeZqqLbbU1Yn"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/o3plWGRcJg0DOsZghC",
+        "id" : "https://www.example.com/OY8batDu5ndLhUrsk6",
         "labels" : {
-          "nFNqcW7IO1o" : "qunEHcKJffoD"
+          "wAeWWzvElhzbM" : "4f1ZfWBnrS16nAkCk"
         }
       } ],
-      "role" : "Funder",
-      "sequence" : 9,
+      "role" : "Designer",
+      "sequence" : 8,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/f8ECwDhx2nRNeL",
-        "name" : "GKdWIUelLkDQ5",
+        "id" : "https://www.example.com/6896kEQPrJCTBGhz7eM",
+        "name" : "aogT4PbObNxw0hxHlU",
         "nameType" : "Personal",
-        "orcId" : "OFoazW0vljU",
-        "arpId" : "7GWznjr3D6lJuDWj7X"
+        "orcId" : "GFNcPSvnHOES2YzLjR",
+        "arpId" : "HcnjnhRoDo"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/dqdDwaiqHlqSizAxvVY",
+        "id" : "https://www.example.com/F5NVPJuULG9rQRcCQC",
         "labels" : {
-          "QvbH8C5d1WzJwqSIO" : "ivHRceRCeTu4iHGgfo"
+          "378APD56O51MYgk0Yu1" : "7EUotHpK1Jq"
         }
       } ],
-      "role" : "CuratorOrganizer",
-      "sequence" : 1,
+      "role" : "Designer",
+      "sequence" : 4,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "W5X81HuLYD3X019Egb",
-    "tags" : [ "emrZmdZo8D23kyy" ],
-    "description" : "e7o566onqfEDeUXu",
+    "npiSubjectHeading" : "xUmgNjGsHWIqs5GO",
+    "tags" : [ "HytkBzsBOFZAZ8" ],
+    "description" : "vs8BbUjfPs22Xdm1",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/xne4kksud7Df4Z"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/4bd4Fdf42ub0W"
       },
-      "doi" : "https://www.example.com/HpaireabR9whxJ9Dxz",
+      "doi" : "https://www.example.com/bIEUQfEFREkfC",
       "publicationInstance" : {
         "type" : "JournalInterview",
-        "volume" : "niKThwHDDzwZv",
-        "issue" : "7FS0aLqNJr5U",
-        "articleNumber" : "emurFuFdnldfTsU2Qce",
+        "volume" : "1S9a2QswkdjmdkOnDzA",
+        "issue" : "WY4g5zyRrelN",
+        "articleNumber" : "dV4KlwYUoYiRxRr",
         "pages" : {
           "type" : "Range",
-          "begin" : "XqBxdYDlqNNzcxjTiwd",
-          "end" : "BuZ3xM7UbmmGoMjv"
+          "begin" : "HBxcaIA8JDj",
+          "end" : "B7UYzN0ouyxxV"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/KFzxT0mbrYJF9RVuuHk",
-    "abstract" : "t2BGh6GTkTYie8jFs"
+    "metadataSource" : "https://www.example.com/qkXiVcNoRdlQ9K23x0N",
+    "abstract" : "ZXv3XepHIRn7"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "cUSSMihst6lg",
-      "mimeType" : "qKPLWi6BeAsQY6",
-      "size" : 2090434302,
+      "name" : "K1qhj7UgA7jNLtzp0qR",
+      "mimeType" : "etl5sSw10urGo3",
+      "size" : 1744337781,
       "license" : {
         "type" : "License",
-        "identifier" : "pSdaQrpYfPSbWN",
+        "identifier" : "AeP6s30uivioWSc0H9",
         "labels" : {
-          "3AcQDlaKHNghuDV4" : "RjUc2IdEqsv"
+          "1ftc1EVRUG" : "5KPRGv1E4WnXCx7niX"
         },
-        "link" : "https://www.example.com/CZRO0AsxJk"
+        "link" : "https://www.example.com/0bQybvSjJQNJVlKJ"
       },
       "administrativeAgreement" : false,
       "publisherAuthority" : false,
@@ -136,19 +136,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/laudantiumsed",
-    "name" : "zz0Uu2xI24BJwLgL6R",
+    "id" : "https://www.example.org/laborumet",
+    "name" : "aX5ggQNxCzx",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "UQGYYK3Isg1b0PgxWnI",
-      "id" : "IVJPHxjZ13fKo9m5jT"
+      "source" : "WWREtcUk606BDwUGJQ",
+      "id" : "PAXcuQlyIRsc"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "JfwdHafPl3mZcyN"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "APPLIED",
+      "applicationCode" : "qIDgihovLvoQIBhgQW"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/maioressaepe" ],
+  "subjects" : [ "https://www.example.org/veroqui" ],
   "modelVersion" : "0.14.24"
 }
