+++ documentation/JournalLeader.json
@@ -1,154 +1,154 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
-  "owner" : "pnXtigBj8Ri6",
+  "status" : "DRAFT_FOR_DELETION",
+  "owner" : "Gc1MUx36CYuQN9Z",
   "resourceOwner" : {
-    "owner" : "pnXtigBj8Ri6",
-    "ownerAffiliation" : "https://www.example.org/esttemporibus"
+    "owner" : "Gc1MUx36CYuQN9Z",
+    "ownerAffiliation" : "https://www.example.org/quiasunt"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/aperiamtemporibus",
+    "id" : "https://www.example.org/quiconsequatur",
     "labels" : {
-      "RfufvWHvsPdcbPn" : "wYDqIpot6YwtWesD"
+      "wOKYEif5WE5R0l0A" : "1ooFaBnbxTMqzAWQx"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/quiqui",
-  "doi" : "https://doi.org/10.1234/facere",
+  "handle" : "https://www.example.org/sapientecorrupti",
+  "doi" : "https://doi.org/10.1234/maiores",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REJECTED",
+    "status" : "APPROVED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "EMa57SDJUL",
-      "author" : "5NJOekXTKDQIUjXnDdY",
+      "text" : "nv92Qc67GJM9PTg",
+      "author" : "CWp5UiGd8IPC0QyF",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/veniamreiciendis",
+  "link" : "https://www.example.org/placeatexcepturi",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "SeeiJnoKqu",
+    "mainTitle" : "muDFCuQTjlmohn",
     "alternativeTitles" : {
-      "bg" : "5MPLwE8ZEGjWTj"
+      "sv" : "BUMwRG8zef6cI6jRfg"
     },
-    "language" : "http://lexvo.org/id/iso639-3/car",
+    "language" : "http://lexvo.org/id/iso639-3/deu",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "eASSeaj79wNA",
-      "month" : "WPON7nqbG1snE6",
-      "day" : "WoCynWsaLz"
+      "year" : "LxSGQ3tsKyQqUYEL0R",
+      "month" : "6F5JTxpuIFg",
+      "day" : "ht6qPFYesoSkHj"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/jtCSqTKh6BPTqM",
-        "name" : "ydk8I3xZOJtFlympeX",
+        "id" : "https://www.example.com/PuSmv5PcVOs3Qv8EZ",
+        "name" : "gpC3U9AfX3vLZy",
         "nameType" : "Organizational",
-        "orcId" : "JbdvuYERKokLU0yLBc",
-        "arpId" : "XvbWGT262g3B0XrMlcY"
+        "orcId" : "kuDztawJ3eLaOz7bjS",
+        "arpId" : "xchTrzaVGg"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/wXp4LOREajbJaRQ",
+        "id" : "https://www.example.com/4uyfo9bfTAqFKW9p",
         "labels" : {
-          "rNL0XTJuF7JAGG" : "rcfC8i7Bp5uz"
+          "UUbE8ZIu1ey0y" : "UJbNukUA4Pk5Eka2iDB"
         }
       } ],
-      "role" : "ResearchGroup",
-      "sequence" : 9,
+      "role" : "Researcher",
+      "sequence" : 5,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/KdPuKOPNMoC",
-        "name" : "b3DxanqwK0",
-        "nameType" : "Organizational",
-        "orcId" : "15FGzSLqladiyo",
-        "arpId" : "CDYwQWhEXdwlY0wd"
+        "id" : "https://www.example.com/HdFyuyspUIugQCiJ",
+        "name" : "AoKgFWN84I8",
+        "nameType" : "Personal",
+        "orcId" : "7mS2SZ4pIahwS",
+        "arpId" : "3LCIYyMhwq4hnVS4G7"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/qR5hq1bbCJWABY6WlH",
+        "id" : "https://www.example.com/Qw9VKOQVzVAN",
         "labels" : {
-          "bfl66EiS3m6RPN8" : "7zAQvY5esmzvF"
+          "RSeXQnHF4WrEt29IDL" : "d3fKGSQXJvD"
         }
       } ],
-      "role" : "WorkPackageLeader",
-      "sequence" : 3,
+      "role" : "ProjectManager",
+      "sequence" : 7,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "6zSfnG9MzoSfn28dfIF",
-    "tags" : [ "sPejiRNZ6EJ2" ],
-    "description" : "IApZu35YCVS",
+    "npiSubjectHeading" : "i52CkpaTkeQ7oAuMcQ",
+    "tags" : [ "xCi5MHpn0IZQM6" ],
+    "description" : "lAhswMVpjYR4U",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/BZLRnMke5yhi7j2vz"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/zxGnUfGyAM08d"
       },
-      "doi" : "https://www.example.com/0pTib7dMks6P91azKO",
+      "doi" : "https://www.example.com/YHz5vAvdsWslHkN24id",
       "publicationInstance" : {
         "type" : "JournalLeader",
-        "volume" : "C3qWHkwT01",
-        "issue" : "rBZMhUQpSX9LsGUCTZG",
-        "articleNumber" : "h4UXZNSeewf",
+        "volume" : "gIG1plCqks0xSim",
+        "issue" : "YakQdNrQ4Zp7f1BBCDv",
+        "articleNumber" : "e6jOxg6Y7jrGzTMTv",
         "pages" : {
           "type" : "Range",
-          "begin" : "W505NmXm3a",
-          "end" : "LnxXI9TztHz35IUy9zT"
+          "begin" : "1Pq6niGwKbp",
+          "end" : "1gr1l1LoSTUg3Mrfx1"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/qwqZqjAoXh95W8TAi",
-    "abstract" : "5jUOkWOjy7qvAp35"
+    "metadataSource" : "https://www.example.com/lD3N7KHVzZg",
+    "abstract" : "c1FFCeaf2lnBxg8y"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "Upa5NGUXHXEhtn",
-      "mimeType" : "bXho6MAyjOVkT",
-      "size" : 1087438778,
+      "name" : "psAMihAjVDE",
+      "mimeType" : "0RP4CWcDO23aJ",
+      "size" : 2066917401,
       "license" : {
         "type" : "License",
-        "identifier" : "siGxJ5zut7BvGVs",
+        "identifier" : "R0a9cbKLskBQluzHmm",
         "labels" : {
-          "fe59sqWrWjI" : "7Q1if3kv8xnOEY9"
+          "Z3HoWj3KuP4IXA" : "mQz0lnn2Li"
         },
-        "link" : "https://www.example.com/wmRTYaV7vVJpW2"
+        "link" : "https://www.example.com/D8vYjA3mft"
       },
-      "administrativeAgreement" : true,
+      "administrativeAgreement" : false,
       "publisherAuthority" : false,
       "embargoDate" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/asperioresinventore",
-    "name" : "dcHxGFGzZbiPOvH",
+    "id" : "https://www.example.org/quiaesse",
+    "name" : "Dm6PGlHKH7",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "eokrJbZPm3LJ",
-      "id" : "JC1zv9t1mYcw"
+      "source" : "arZSr7bgsmCQq81fSf",
+      "id" : "d6y06gIhZXdeX"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
-      "approvalStatus" : "DECLINED",
-      "applicationCode" : "6rIuiNrEAr77y"
+      "approvedBy" : "NMA",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "SIxdj210ASkR"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/quiofficiis" ],
+  "subjects" : [ "https://www.example.org/errorunde" ],
   "modelVersion" : "0.14.24"
 }
