+++ documentation/JournalLetter.json
@@ -1,25 +1,25 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
-  "owner" : "8xr6GTGQJUeZK78Nz4",
+  "status" : "DRAFT_FOR_DELETION",
+  "owner" : "hviTRqLxBcpi",
   "resourceOwner" : {
-    "owner" : "8xr6GTGQJUeZK78Nz4",
-    "ownerAffiliation" : "https://www.example.org/exercitationemesse"
+    "owner" : "hviTRqLxBcpi",
+    "ownerAffiliation" : "https://www.example.org/eumnon"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/molestiaeautem",
+    "id" : "https://www.example.org/autamet",
     "labels" : {
-      "sAxpDwVq54k4NCoi" : "xEVy3Lgft8"
+      "A3eI0pMazOjhK" : "BvjTPxLSEDbXZAurrz"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/utodio",
-  "doi" : "https://doi.org/10.1234/quaerat",
+  "handle" : "https://www.example.org/accusamuseum",
+  "doi" : "https://doi.org/10.1234/unde",
   "doiRequest" : {
     "type" : "DoiRequest",
     "status" : "REQUESTED",
@@ -27,107 +27,107 @@
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "wpGcBbPlOk5SwQ",
-      "author" : "7yLfq87sFDrAO",
+      "text" : "DHoK3R8903KDmyz",
+      "author" : "PvRaokYUr1dVg",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/oditad",
+  "link" : "https://www.example.org/nesciuntest",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "XLbAITSBeAYc",
+    "mainTitle" : "MudXwgexFfrklK",
     "alternativeTitles" : {
-      "fi" : "ii0NiLi2jZ0OI0wVxS"
+      "zh" : "4h03aHqxSiZkGUbF"
     },
-    "language" : "http://lexvo.org/id/iso639-3/isl",
+    "language" : "http://lexvo.org/id/iso639-3/nor",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "kUCzWuiW3kln4X",
-      "month" : "agct4UOuIwG57XSqZ3",
-      "day" : "dydvpf8ugHSw977"
+      "year" : "E8vgPMo2BOkndCbABtc",
+      "month" : "MoYXWWGQbQ0JR",
+      "day" : "LhQUePnjBBXhiW"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/QUe23A9TUtNSC6HO",
-        "name" : "7tunyjRp28s",
+        "id" : "https://www.example.com/IYuLnrj8aRgyK",
+        "name" : "TtZ8dy7TIAMR",
         "nameType" : "Personal",
-        "orcId" : "fBhMMdBJa78W",
-        "arpId" : "AxIrz2tOv0yo0z4Xbbt"
+        "orcId" : "UdaKIQTCJYX",
+        "arpId" : "bMMGDspyQRM12EB"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/izrY910GL6xJw",
+        "id" : "https://www.example.com/nZC2CS3eObN",
         "labels" : {
-          "O6ruHlTgxW" : "EmRWh3jgpbyuLDc"
+          "pfCzYzB5jue" : "bLR3IjWgRUkAYUePp7H"
         }
       } ],
-      "role" : "Supervisor",
+      "role" : "Designer",
       "sequence" : 0,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Zs9Xd4rqnNDqw7WgJA",
-        "name" : "4KdwLA9Y0othfLpRa6J",
-        "nameType" : "Personal",
-        "orcId" : "LtsQz4yMlp",
-        "arpId" : "ICpdfhlwbmNED0T3Fb5"
+        "id" : "https://www.example.com/oBGeM8li8Y",
+        "name" : "7Gn0AFuLWEL",
+        "nameType" : "Organizational",
+        "orcId" : "LseLSV9GEqK",
+        "arpId" : "QXYdiQJVSqahvE5jd4"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/G0tsdRpJbFQqQk5D",
+        "id" : "https://www.example.com/RR5wtYD5w8fu4",
         "labels" : {
-          "uG9UJyt0un" : "FFlyXoOSB2w"
+          "H3X87aSYYA9hVCk" : "rh3GNkN3C2euspUB2F"
         }
       } ],
-      "role" : "EditorialBoardMember",
-      "sequence" : 9,
+      "role" : "RegistrationAgency",
+      "sequence" : 0,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "uZ02TNl2S2ICpKV",
-    "tags" : [ "jEsp7bcbPviKF" ],
-    "description" : "YaoomQE26Piea80W0fA",
+    "npiSubjectHeading" : "91lOKtPV0DegChX8NGA",
+    "tags" : [ "LIdivnKQ8uoYxLWtd3I" ],
+    "description" : "sUdr59fJKkWCmN4CA",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/R0PFphFnHU"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/RORi6ireTa3OqyVklK"
       },
-      "doi" : "https://www.example.com/9j3Lh38YtoAI",
+      "doi" : "https://www.example.com/PlaHSDIJTK0VplLgVI",
       "publicationInstance" : {
         "type" : "JournalLetter",
-        "volume" : "zhYJLvEnt3s",
-        "issue" : "IJzDJe9lBfVT12TVG",
-        "articleNumber" : "E2eVLJBZMiQ6UIeY",
+        "volume" : "PZXH45o3A7H7hG0DTu",
+        "issue" : "eICUCbVbeL",
+        "articleNumber" : "WOXF2H1O3S1",
         "pages" : {
           "type" : "Range",
-          "begin" : "2KE7m6uTETn2",
-          "end" : "I1gETUfhI6xgQd1wohM"
+          "begin" : "fsTPwKP4wsyF",
+          "end" : "JVl6KJ4rdz3NLo"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/h8BH5JpsIFxyKD",
-    "abstract" : "WpgwF2aOHTjTvA2"
+    "metadataSource" : "https://www.example.com/Zd3vMHaxtuYqhIt0",
+    "abstract" : "T0JVq0m6GXFpg0q"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "HVPyeEZ4wnqP9xt5k",
-      "mimeType" : "ySTRuW7TBEUwa1BP",
-      "size" : 1686431076,
+      "name" : "3qZ7DoZ8ycCWjADIdYX",
+      "mimeType" : "khYzUMkPHR",
+      "size" : 1204857731,
       "license" : {
         "type" : "License",
-        "identifier" : "m3sGfbgVUwVQBL",
+        "identifier" : "RgwOyLpWKeMkJHTN",
         "labels" : {
-          "9mk6w1D7QMuqz56" : "fZtVteJXNRiLuXqXi"
+          "WQisFZONUw2KFe" : "a8fSwaI95XQ7sGzGYeE"
         },
-        "link" : "https://www.example.com/ASovQtGX9V7S"
+        "link" : "https://www.example.com/qylvxgdFbjvEf8hDzxl"
       },
       "administrativeAgreement" : true,
       "publisherAuthority" : false,
@@ -136,19 +136,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/etmollitia",
-    "name" : "U0wNXQvF6ogEuYyQz",
+    "id" : "https://www.example.org/consequaturqui",
+    "name" : "96GeMjUPyk",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "7yeOHLh1UYlx",
-      "id" : "pdIV6iADmzVIOIIYh7G"
+      "source" : "YLOKe3BEoO",
+      "id" : "RHCD44EURC1C"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "zdqq21sJlGr"
+      "approvedBy" : "REK",
+      "approvalStatus" : "APPROVED",
+      "applicationCode" : "5JG95xUJTc67wxG"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/eaqueexpedita" ],
+  "subjects" : [ "https://www.example.org/voluptateminventore" ],
   "modelVersion" : "0.14.24"
 }
