+++ documentation/JournalReview.json
@@ -2,153 +2,153 @@
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "PUBLISHED",
-  "owner" : "l4HrIBkczV",
+  "owner" : "mr6qm3M7jaMmOPeS",
   "resourceOwner" : {
-    "owner" : "l4HrIBkczV",
-    "ownerAffiliation" : "https://www.example.org/consequatureos"
+    "owner" : "mr6qm3M7jaMmOPeS",
+    "ownerAffiliation" : "https://www.example.org/eiusreprehenderit"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/quiautem",
+    "id" : "https://www.example.org/utrerum",
     "labels" : {
-      "pUzErsdZ1uD" : "5kXbkduWixf0y86ww"
+      "GnXEsEMBzf4k1u" : "IFlrp18yafC"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/praesentiumaliquam",
-  "doi" : "https://doi.org/10.1234/deleniti",
+  "handle" : "https://www.example.org/etquis",
+  "doi" : "https://doi.org/10.1234/a",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REQUESTED",
+    "status" : "APPROVED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "z2Bj7FD9VwyE7FG",
-      "author" : "UFYUdHVqnyqFgPUzK",
+      "text" : "d6RJ9f7twdTdoP9",
+      "author" : "5r5YThKXvNeopxe6H",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/optiodolores",
+  "link" : "https://www.example.org/consequatura",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "uEwU7aZubNpDNaCYp",
+    "mainTitle" : "bFY3eQoNPzYalpudL1",
     "alternativeTitles" : {
-      "en" : "9Djx075abrqy"
+      "en" : "Qr3wAQJO2Gaj"
     },
-    "language" : "http://lexvo.org/id/iso639-3/fin",
+    "language" : "http://lexvo.org/id/iso639-3/nob",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "5nkVyp3rM2",
-      "month" : "SvWbnP6EB0bYmasy",
-      "day" : "rrtBK1P66o4R2wjxoJP"
+      "year" : "rbBooQojFzAGCK6nFc",
+      "month" : "0yxOERzKMH8S",
+      "day" : "g09LGIgbwAhSLXwP"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/R89QNfn6DVX",
-        "name" : "s3uqiC9kpm0VqYh0tP2",
-        "nameType" : "Organizational",
-        "orcId" : "eQnWikDsXu8AyCL2Fo",
-        "arpId" : "baSj2Kv4SBIurbay7y"
+        "id" : "https://www.example.com/ZgeGXvsd1MgXWDh8FP",
+        "name" : "iNlrDn4KaMbg",
+        "nameType" : "Personal",
+        "orcId" : "Tu2z6FjivcGpvn5",
+        "arpId" : "GC209BJ7Wxj6x7Zf9"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/xJTseWydXslu",
+        "id" : "https://www.example.com/mTJRDPwHpBC1g6e0wh",
         "labels" : {
-          "aErrUm8v8U9OfrgNpm" : "XulrS9f1DsDniv5BhKE"
+          "DOxa3Bf2U2ixhEWy" : "bQJjMzkpnw31aygG4"
         }
       } ],
-      "role" : "ProjectMember",
-      "sequence" : 5,
+      "role" : "Producer",
+      "sequence" : 2,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/w9FgnPJzPCR4tDqBgj",
-        "name" : "sPm4j4Ql8gnVsJ0",
-        "nameType" : "Organizational",
-        "orcId" : "ALKRNywAI7U",
-        "arpId" : "JolavU2pOkiUYEPn"
+        "id" : "https://www.example.com/3tcRIByq3Bg",
+        "name" : "5cFSlCvm9Zt",
+        "nameType" : "Personal",
+        "orcId" : "jLxeS0L01a9mVsS",
+        "arpId" : "N8RmamE1OA7Pnz"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/OWdW4NMcLEqW",
+        "id" : "https://www.example.com/Z0nWW3vbwZ",
         "labels" : {
-          "zHsOoFJHO5L" : "gCg1w5QBuUU"
+          "2bUVECxJ6yRQ" : "e4NoULSaKJ8L7E91VFD"
         }
       } ],
-      "role" : "RegistrationAgency",
-      "sequence" : 3,
+      "role" : "RegistrationAuthority",
+      "sequence" : 5,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "YEsDQBnD798XJqfn",
-    "tags" : [ "gjnneIvLOHMZr0" ],
-    "description" : "ITpIu3dHgh5CM8Xh8g",
+    "npiSubjectHeading" : "cMahnx7xOrno",
+    "tags" : [ "kRAbuE0s73gv9iZ" ],
+    "description" : "wl1HImvuflT",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/aBmBwxca697vS9U"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/xFbsY8ET5W8F"
       },
-      "doi" : "https://www.example.com/Eh5oBk1Wkp",
+      "doi" : "https://www.example.com/046SmNsnT1D",
       "publicationInstance" : {
         "type" : "JournalReview",
-        "volume" : "7cLwhklemG2stctoKi",
-        "issue" : "mACVtt9a172i1LAU261",
-        "articleNumber" : "nZ4qvPEMT00Pxw",
+        "volume" : "Es7ThH5ifIG",
+        "issue" : "7xI6CGpNFOFFGw",
+        "articleNumber" : "G6hDw0yohCwyPaW",
         "pages" : {
           "type" : "Range",
-          "begin" : "PIsGGWNCT2MB",
-          "end" : "YKfu0ZnXaahw7"
+          "begin" : "cMbB5wfFBywvML",
+          "end" : "V4m4WfWydbE6aRZMWfk"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/uXuLlvWNojce86ou7ks",
-    "abstract" : "2biGOPtTze"
+    "metadataSource" : "https://www.example.com/k75iSxhTORwsqDvDi",
+    "abstract" : "ustq6qnNcEV"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "O3ZgUPwHs2XFY29",
-      "mimeType" : "kRuH2RbRTyHlBeLaD",
-      "size" : 1854101787,
+      "name" : "c8tgJEth1t9EpQqaIWF",
+      "mimeType" : "PxnqqxxCMyJSF",
+      "size" : 510755112,
       "license" : {
         "type" : "License",
-        "identifier" : "cHbwEQXGtKwA6o",
+        "identifier" : "SthSoSwslcsLP",
         "labels" : {
-          "B3CtAj6LIWlgv81vi3u" : "T6Dzt3dYPfZd0rbSEQx"
+          "2KAniFgR6iJWJ" : "6qHL1fWaYdzW"
         },
-        "link" : "https://www.example.com/S6j7QUtq4bWYxmdfcXL"
+        "link" : "https://www.example.com/5ZVYjuWpfx"
       },
-      "administrativeAgreement" : true,
+      "administrativeAgreement" : false,
       "publisherAuthority" : false,
       "embargoDate" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/totamest",
-    "name" : "hE2Bre6W6ikgh6",
+    "id" : "https://www.example.org/siterror",
+    "name" : "K9HMI5c0Pk5T1oRpg3h",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "NWNfUFLvSooEBnr",
-      "id" : "N1yC0FMUD6kRuCpNS40"
+      "source" : "GRQD7ZoFeqC",
+      "id" : "9ZKMNCqALIms"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "APPROVED",
-      "applicationCode" : "c6qO19PwNOYNJ3t"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "ikwXGrFCcmEs6ArMYhO"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/itaqueimpedit" ],
+  "subjects" : [ "https://www.example.org/accusantiumdoloremque" ],
   "modelVersion" : "0.14.24"
 }
