+++ documentation/JournalShortCommunication.json
@@ -1,154 +1,154 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
-  "owner" : "Eb3W6fYm8m66pNIJ",
+  "status" : "NEW",
+  "owner" : "YqCi7y7TNWlCJps",
   "resourceOwner" : {
-    "owner" : "Eb3W6fYm8m66pNIJ",
-    "ownerAffiliation" : "https://www.example.org/utsaepe"
+    "owner" : "YqCi7y7TNWlCJps",
+    "ownerAffiliation" : "https://www.example.org/cumarchitecto"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/voluptatelaborum",
+    "id" : "https://www.example.org/etplaceat",
     "labels" : {
-      "lXdsqPHiALgyDx" : "6yOEvkfKcY3heYhH5"
+      "9LAx1FD4qJq5uPzG6" : "WqigJzLlDy2vaic7M1"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/aliquamdolorem",
-  "doi" : "https://doi.org/10.1234/facilis",
+  "handle" : "https://www.example.org/maximeexpedita",
+  "doi" : "https://doi.org/10.1234/provident",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REJECTED",
+    "status" : "APPROVED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "8Rxy7qnbRNyQB90",
-      "author" : "R3iEwSBOD3aenxHvYsr",
+      "text" : "6QlY88X9AvKk",
+      "author" : "b7u6vj1yF6",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/quiadolorem",
+  "link" : "https://www.example.org/fugamagni",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "QIAEEWVNDR",
+    "mainTitle" : "Xm8Nz1O01wax",
     "alternativeTitles" : {
-      "nb" : "SjvTnNPVTzETY"
+      "en" : "TAwqCcI1eTVVc"
     },
-    "language" : "http://lexvo.org/id/iso639-3/deu",
+    "language" : "http://lexvo.org/id/iso639-3/spa",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "TPOJaclESLa1sYz63d",
-      "month" : "YAWz2rporTtDzkT",
-      "day" : "qiZjka05Wq"
+      "year" : "ROj0yTJAxNf0hTU",
+      "month" : "9wJBv593lGm",
+      "day" : "3QQdUly29j00sdRSq"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/lRy9iRVJEHTbKLzmR7",
-        "name" : "Oh0RxCgpUsY",
-        "nameType" : "Organizational",
-        "orcId" : "sVUx0izBuQXOUi0",
-        "arpId" : "OGZnJqIJc3kXc1td"
+        "id" : "https://www.example.com/5ToMdBpVRVo7pp0n",
+        "name" : "fiqeQcncoFbZN7ysGU",
+        "nameType" : "Personal",
+        "orcId" : "DSie4Dt0j6cvkYUT4Je",
+        "arpId" : "jGeXJ6KIeVCk7h4a"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/JTfZHsL7J0ggkRnJT",
+        "id" : "https://www.example.com/ZW2nLVmCiezWKpl",
         "labels" : {
-          "R6M0iPyXlenjwI7ot0" : "hayfxFTac3"
+          "NslwnrBYTi" : "g2stDK0ZSEojG"
         }
       } ],
-      "role" : "Designer",
-      "sequence" : 7,
+      "role" : "Illustrator",
+      "sequence" : 1,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/rkZwMwoZjcdgHFKq92l",
-        "name" : "1XSJXX3bbLPv1fkx",
-        "nameType" : "Organizational",
-        "orcId" : "GzNHOqppSit7",
-        "arpId" : "WtfDU9RVTEyRN"
+        "id" : "https://www.example.com/xAaKNS2CkQipvkTIOz2",
+        "name" : "njhGSfG7z8Zxj",
+        "nameType" : "Personal",
+        "orcId" : "Ls8u5GGHQtw1GIOV",
+        "arpId" : "oDhQJrV7xVmNhH"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/1Djj5Atarh",
+        "id" : "https://www.example.com/Aj96f05DHeYu",
         "labels" : {
-          "HZjQH94mNxKlcGQofe6" : "3FwU1eaLcr"
+          "8K0KLCklENA" : "I5N90b8LY9mIg"
         }
       } ],
       "role" : "WorkPackageLeader",
-      "sequence" : 6,
+      "sequence" : 7,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "3xmB9l5dm7gKTFBNy",
-    "tags" : [ "UYfrLSYvpNf8" ],
-    "description" : "ZR2hB2hwPc",
+    "npiSubjectHeading" : "vCG9KB7nNbG8nI",
+    "tags" : [ "XfswQX7IRfiKgVLC" ],
+    "description" : "gKsCYMTnwuDz6RV6",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/eRd7yZDnY2j"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/9LfOZPaW8f"
       },
-      "doi" : "https://www.example.com/tGUq1g30gcpBBMHRB",
+      "doi" : "https://www.example.com/qcELn4yC623P",
       "publicationInstance" : {
         "type" : "JournalShortCommunication",
-        "volume" : "WGNdcptlEKuTWCeFq",
-        "issue" : "UWuPbtImglGCMb5",
-        "articleNumber" : "peSKT1jZKmfvXjqXR2",
+        "volume" : "h2zDshNehJGJTBmzv",
+        "issue" : "VwZDN7ZREJZd",
+        "articleNumber" : "ILdXMTOZJvQ",
         "pages" : {
           "type" : "Range",
-          "begin" : "4TI6SNpI3I",
-          "end" : "FJXfAhsiWdIA"
+          "begin" : "Zq8Bzt0JdIS69Me",
+          "end" : "Hd4Sp04HRXnJDOq2Lu"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/Wkt8R2aWKd4u",
-    "abstract" : "iXWozLgzaUUvSs2s2"
+    "metadataSource" : "https://www.example.com/F5dyEwQkTlUl5q",
+    "abstract" : "GLHPRhRhr2J21liT4"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "cu1ERBetFTl",
-      "mimeType" : "riZHPRhj0RKpG0T4c",
-      "size" : 1144134008,
+      "name" : "HGMB5doH0P0sBe",
+      "mimeType" : "wTiwm0YkBfi7uUyn",
+      "size" : 262541493,
       "license" : {
         "type" : "License",
-        "identifier" : "U2qddbo7ngZ17A",
+        "identifier" : "5bP5uB4CTXszdpR",
         "labels" : {
-          "v2cBKQQjsjR0Df" : "QmVXxo3nMjAcZa8bQ0"
+          "vIzYYpy6pSKn9EP" : "fKxH6apxamYPNI9Ud"
         },
-        "link" : "https://www.example.com/DNgKIW3zrKMikQU4"
+        "link" : "https://www.example.com/Eavb53s1A4f4jj3g"
       },
-      "administrativeAgreement" : false,
+      "administrativeAgreement" : true,
       "publisherAuthority" : false,
       "embargoDate" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/voluptatempariatur",
-    "name" : "zNkndJdTmp6bsosd9",
+    "id" : "https://www.example.org/quiad",
+    "name" : "lKOZIHKLrxn",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "o9j5xePEZzWMQzU2t",
-      "id" : "qI1JYZx3Rl"
+      "source" : "SSnvwWxasnBdfIQ",
+      "id" : "txE6WxoYniASJPJaR"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "REJECTION",
-      "applicationCode" : "SAHmGwJ6gRlhYM"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "7vFs1goNAu6rTx"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/molestiaea" ],
+  "subjects" : [ "https://www.example.org/evenietmagni" ],
   "modelVersion" : "0.14.24"
 }
