+++ documentation/Lecture.json
@@ -2,24 +2,24 @@
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "PUBLISHED",
-  "owner" : "Y1j3TblCh3DYGNxIPA",
+  "owner" : "WGUVpk1vsd1WcfEwOO",
   "resourceOwner" : {
-    "owner" : "Y1j3TblCh3DYGNxIPA",
-    "ownerAffiliation" : "https://www.example.org/ettempora"
+    "owner" : "WGUVpk1vsd1WcfEwOO",
+    "ownerAffiliation" : "https://www.example.org/illoipsam"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/abomnis",
+    "id" : "https://www.example.org/consequaturdolores",
     "labels" : {
-      "9xJ21P5X1oRPL8uV" : "KYPZ3crmkSmiS"
+      "8youneejpQ5fm" : "I8kbVvJ2byGv"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/ducimusqui",
-  "doi" : "https://doi.org/10.1234/vel",
+  "handle" : "https://www.example.org/occaecatibeatae",
+  "doi" : "https://doi.org/10.1234/magni",
   "doiRequest" : {
     "type" : "DoiRequest",
     "status" : "APPROVED",
@@ -27,94 +27,93 @@
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "ppIx2WlImmqCJDc",
-      "author" : "2Xb73yK2SU",
+      "text" : "0MAziwLcgXN7A",
+      "author" : "R01z9TeGIdRyTR2Mdt",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/quosatque",
+  "link" : "https://www.example.org/quoin",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "QnWPOPl0c2Y",
+    "mainTitle" : "RybHetWEDAJndI",
     "alternativeTitles" : {
-      "af" : "MpfUO92x1dlh"
+      "sv" : "ZFfZAbmUw45oCVWvj"
     },
-    "language" : "http://lexvo.org/id/iso639-3/dan",
+    "language" : "http://lexvo.org/id/iso639-3/fra",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "M25fptRclQ",
-      "month" : "hLxil6H5q7geL",
-      "day" : "7Xpzb4FiYEdd"
+      "year" : "AhM6HtSVioLE9q",
+      "month" : "cyjwDBkVhpiL9uRuSN",
+      "day" : "QqBsN4RtD9HkCq1zxcm"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/McWfkUrDot",
-        "name" : "f16aaMWEsSLuai",
+        "id" : "https://www.example.com/7oxT1QZMA3RZGJ",
+        "name" : "he2JaH0To2CnM",
         "nameType" : "Organizational",
-        "orcId" : "eHtCqRznFsNPDfTAfN",
-        "arpId" : "VBSJJYvnNV10"
+        "orcId" : "95VGmJ7nxOtjIXIK1Q",
+        "arpId" : "Fltpumf1RjFt8IJOQ"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/aCgL1noWxVYj2bcWkP",
+        "id" : "https://www.example.com/H44t8GQ9NYf",
         "labels" : {
-          "8ojqBA2Ec9" : "BSnRY2Jw4gsSbmMvUl"
+          "ZO5axdNAIn" : "9WkbI3k56EPVo"
         }
       } ],
-      "role" : "Funder",
-      "sequence" : 5,
+      "role" : "RegistrationAgency",
+      "sequence" : 9,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/VOpsfr9v1u",
-        "name" : "gzsPG3LHX3sQtU4Z",
-        "nameType" : "Organizational",
-        "orcId" : "rbl5DM61qPuB",
-        "arpId" : "DTYQuiOgUV"
+        "id" : "https://www.example.com/Cz5isQoPrd",
+        "name" : "qcvCAFyGEJPgR",
+        "nameType" : "Personal",
+        "orcId" : "7fPyvP1PUiyyujzGrsH",
+        "arpId" : "yoYnlujbpwnDFQ"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/iN2pFIiKBmi02",
+        "id" : "https://www.example.com/8CO4orwFYle",
         "labels" : {
-          "f5CADBhVuNBOoeM" : "bif7rRDcrPq7u"
+          "Ei6pMZntD7hSAFg" : "3uLK5HKlOl"
         }
       } ],
-      "role" : "ProgrammeLeader",
-      "sequence" : 0,
+      "role" : "ProgrammeParticipant",
+      "sequence" : 6,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "mDdo5LhP2V1",
-    "tags" : [ "NJfE05bNs6T7" ],
-    "description" : "lnovVgC4DzXNY",
+    "npiSubjectHeading" : "aVTEuBVWH7huuCrCd",
+    "tags" : [ "AcPJP1HXPg5AVcKA" ],
+    "description" : "VXIbwSaaqcWPeCs4rwV",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Event",
-        "label" : "Gef5tTzVFsvf",
+        "label" : "UMpEVtjT9BOe4",
         "place" : {
           "type" : "UnconfirmedPlace",
-          "label" : "jUnW4lhNNih",
-          "country" : "VqpysR1aZBgW0tTqh"
+          "label" : "KOcwKZi7y5mc",
+          "country" : "WJRzG1xIlzBa"
         },
         "time" : {
-          "type" : "Period",
-          "from" : "2020-09-23T09:51:23.044996Z",
-          "to" : "2020-09-23T09:51:23.044996Z"
+          "type" : "Instant",
+          "value" : "2020-09-23T09:51:23.044996Z"
         },
         "agent" : {
           "type" : "Organization",
-          "id" : "https://www.example.com/18OJnVTVu76YkscElbo",
+          "id" : "https://www.example.com/h5yyvELAbguQmgxD",
           "labels" : {
-            "v4yDgbFV8akGGNkD4hr" : "Xobuel8Bc016"
+            "iih9heNEadUsZEq3" : "1ZfumqfESBxSKhWA"
           }
         },
-        "product" : "https://www.example.com/npEwRPVYSJt5YX9"
+        "product" : "https://www.example.com/ryX3mXIiO23Q"
       },
-      "doi" : "https://www.example.com/7XbZ8jim7N96MS",
+      "doi" : "https://www.example.com/vaJ9HbhpwuZM",
       "publicationInstance" : {
         "type" : "Lecture",
         "peerReviewed" : false,
@@ -123,45 +122,45 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/tKRHEt7nUK3cjb",
-    "abstract" : "4JXiquLEwT7E1CcNt"
+    "metadataSource" : "https://www.example.com/158QQ5POSUzfrcNuch",
+    "abstract" : "mrrC1iS1vTMTrZ34rv6"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "Whr3RzEK3szgo8yrwE",
-      "mimeType" : "qsZxMI3gjr",
-      "size" : 1868083796,
+      "name" : "IAt4Ken7TzK6",
+      "mimeType" : "RcRrzsWjO7OcHW1rlF",
+      "size" : 161057905,
       "license" : {
         "type" : "License",
-        "identifier" : "l5vJ2xurUqMFB",
+        "identifier" : "DcrwFEwq5EH5TxAC",
         "labels" : {
-          "ytOOdtQv07vSf" : "Spqb7WjUt6"
+          "4H7DQ5jZNF" : "UBNlozUuyS07Xynx24J"
         },
-        "link" : "https://www.example.com/ZDgiqPt6tXI"
+        "link" : "https://www.example.com/9Ugb1sKRUW"
       },
-      "administrativeAgreement" : false,
+      "administrativeAgreement" : true,
       "publisherAuthority" : false,
       "embargoDate" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/earumest",
-    "name" : "VpjLcF6jELLX81",
+    "id" : "https://www.example.org/accusantiumminima",
+    "name" : "eec6w0vOG8qpn",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "9iPmeJk2mznEUPiH8Te",
-      "id" : "oDPaQpePuovNos"
+      "source" : "VImHUykoOi82ILc",
+      "id" : "BaO5wSZvxtfALv7UR"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "REK",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "Zt0uTgoK1t8Igh8zGXv"
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "976JMGktHIJi04AcPsB"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -169,6 +168,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/cumducimus" ],
+  "subjects" : [ "https://www.example.org/porroquis" ],
   "modelVersion" : "0.14.24"
 }
