+++ documentation/OtherPresentation.json
@@ -2,119 +2,118 @@
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "NEW",
-  "owner" : "pKjqK2kqG34Ut",
+  "owner" : "uihk0PwYg73qRcj5",
   "resourceOwner" : {
-    "owner" : "pKjqK2kqG34Ut",
-    "ownerAffiliation" : "https://www.example.org/estad"
+    "owner" : "uihk0PwYg73qRcj5",
+    "ownerAffiliation" : "https://www.example.org/voluptatemquia"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/delenitieaque",
+    "id" : "https://www.example.org/maioresaut",
     "labels" : {
-      "judb3Bt2O8m4MIQa8qr" : "MwB9QPNS6nsXtFuF"
+      "1lYUSWVxUJxD2abF" : "1XNoXnnYb23jD"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/omnisbeatae",
-  "doi" : "https://doi.org/10.1234/dicta",
+  "handle" : "https://www.example.org/optioautem",
+  "doi" : "https://doi.org/10.1234/quia",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "APPROVED",
+    "status" : "REQUESTED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "MeC8UitOm3CsK",
-      "author" : "MxxskIZ23GsllcJQ",
+      "text" : "4ntOL8VLHxG0kr",
+      "author" : "XYTKqBgaLQB4O9",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/laborumsoluta",
+  "link" : "https://www.example.org/easunt",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "3qbDPfsP6kw39R",
+    "mainTitle" : "KglDcplvy8WiC",
     "alternativeTitles" : {
-      "it" : "eCKo5W1EXqWgU"
+      "nl" : "JmwJWx7qVw"
     },
-    "language" : "http://lexvo.org/id/iso639-3/nor",
+    "language" : "http://lexvo.org/id/iso639-3/ita",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "FrbFPZPpT8OlJsyjkJ",
-      "month" : "mMrVWNkvijMQ",
-      "day" : "A8W5fl7NuclOr2xC"
+      "year" : "qQbyadiRlQXgI",
+      "month" : "oC6DLF2HU5vu1B3nkkg",
+      "day" : "dn7uicisxEHJ"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/oFUaKsCBZs",
-        "name" : "62Aga4iyHcO2ckIXa",
+        "id" : "https://www.example.com/ZmbU48Hlu97mW4e2",
+        "name" : "2bD0EoVvi2UvhLrusn",
         "nameType" : "Organizational",
-        "orcId" : "iTpRP7HH026eS3Y1",
-        "arpId" : "pQVeQ29BRgsKp"
+        "orcId" : "AwBSgv5y1StPydec",
+        "arpId" : "wDOBP1ExbCpp"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/URfl0YsVFOnmjboVtUv",
+        "id" : "https://www.example.com/gY9aI8is73NsfE4i0G",
         "labels" : {
-          "DeDDnGyyDjepkSJbU" : "pFbbntWjn4B2vo7EO8"
+          "Wayf3LUibu6CK" : "G3aG6gncsRHXl"
         }
       } ],
-      "role" : "Designer",
-      "sequence" : 0,
+      "role" : "Sponsor",
+      "sequence" : 6,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/7S6gUYMytJZ",
-        "name" : "nuAR2NHqHf",
-        "nameType" : "Personal",
-        "orcId" : "0irbuFSTZTTOcPT",
-        "arpId" : "dgNAwdrn2qi8tp6"
+        "id" : "https://www.example.com/FOp10VPxGKWCrcZrI",
+        "name" : "asTUMWghSA7p1bomM",
+        "nameType" : "Organizational",
+        "orcId" : "l2zXLdjAw8wyWuQhc6Z",
+        "arpId" : "b8Nj2kBNTjbA"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/WoaMJGcEVqG17AON",
+        "id" : "https://www.example.com/100KnPtD1F",
         "labels" : {
-          "c42Fwzz76Qdy6exT" : "o18idslsEsHW3UfYhy"
+          "FVbzbIGm3by" : "SQGhtgpeRCy2nPxJkZ"
         }
       } ],
-      "role" : "Researcher",
-      "sequence" : 3,
+      "role" : "Advisor",
+      "sequence" : 8,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "g78VmAyQDri5G166",
-    "tags" : [ "RHlhEax5kPvfn8OU" ],
-    "description" : "A0l85aCErV8eo2",
+    "npiSubjectHeading" : "Ofa8ygAriQkTi2",
+    "tags" : [ "nBLko58DUzw" ],
+    "description" : "b0izhnsXjU6J0yNjhsv",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Event",
-        "label" : "Q1pUBwzZNOJJmpfNUSP",
+        "label" : "mex9qLZGniO",
         "place" : {
           "type" : "UnconfirmedPlace",
-          "label" : "Mk8KDQTMZdGq1",
-          "country" : "Q1ouymZw6VVU944"
+          "label" : "509XAUJA65sFSIrb9v5",
+          "country" : "PXlDjOvkuqhwN25"
         },
         "time" : {
-          "type" : "Period",
-          "from" : "2020-09-23T09:51:23.044996Z",
-          "to" : "2020-09-23T09:51:23.044996Z"
+          "type" : "Instant",
+          "value" : "2020-09-23T09:51:23.044996Z"
         },
         "agent" : {
           "type" : "Organization",
-          "id" : "https://www.example.com/Mv5ht28qovD",
+          "id" : "https://www.example.com/OTdaw2eJPVO8Np0",
           "labels" : {
-            "GPpLXOqCptk" : "m5unQhx9La"
+            "lbFd1VnEdHiXfLhjH" : "EpRXtqEehJP"
           }
         },
-        "product" : "https://www.example.com/4CQvhXMz4gHZ8oQa"
+        "product" : "https://www.example.com/ENENFQODwKs"
       },
-      "doi" : "https://www.example.com/gNTQ0FQwKCg",
+      "doi" : "https://www.example.com/u3IKcMjICh6",
       "publicationInstance" : {
         "type" : "OtherPresentation",
         "peerReviewed" : false,
@@ -123,24 +122,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/A5VPBrMAWpFV0ui3zo",
-    "abstract" : "SPJq8tIPD0VcvB"
+    "metadataSource" : "https://www.example.com/IJI2AcRBLdQ476i7VX",
+    "abstract" : "fJnO0CAEJe8zaIh"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "gb65S6HNos6A5RiTnJs",
-      "mimeType" : "nbR6ymgfG7XI0Fc",
-      "size" : 1669322651,
+      "name" : "eDZ5IK2IftTYn5",
+      "mimeType" : "J7UY4vIrOxJ",
+      "size" : 565426518,
       "license" : {
         "type" : "License",
-        "identifier" : "dhaOq46ti7vdwnfzta",
+        "identifier" : "4tELZCUkTY96FEYFs",
         "labels" : {
-          "R96XLVNOWruIOq940q" : "TG5JZJdnktVOcPysuI"
+          "aDBg7BnorIWriAxWQR" : "NxZs2DwcOQTKXlJbSJ"
         },
-        "link" : "https://www.example.com/zn1K6lum3kZJ"
+        "link" : "https://www.example.com/Xav2havVeZp"
       },
       "administrativeAgreement" : true,
       "publisherAuthority" : false,
@@ -149,19 +148,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/doloresnostrum",
-    "name" : "6pqneZOUTUm3AbN2uk",
+    "id" : "https://www.example.org/atnam",
+    "name" : "8CIqkDhKmcxVph9tC",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "FISxu7DF3IaGf33W",
-      "id" : "LcvomUUU9jE"
+      "source" : "6Sthc5uDHNNgjDSFm",
+      "id" : "zMU91O4Q6psXRak2"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "IXD54k2s27dPJwtkS6"
+      "approvedBy" : "DIRHEALTH",
+      "approvalStatus" : "APPROVED",
+      "applicationCode" : "fP2BP1jphotBI"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -169,6 +168,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/necessitatibuset" ],
+  "subjects" : [ "https://www.example.org/odithic" ],
   "modelVersion" : "0.14.24"
 }
